# lib/api-zod/src/generated/types/contrato.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ContabDOC API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContratoStatus } from "./contratoStatus";

export interface Contrato {
  id: number;
  clienteId: number;
  numeroContrato: string;
  valorContrato?: string;
  dataContrato?: string;
  diaVencimento?: number;
  dataVencimento?: string;
  objeto?: string;
  status?: ContratoStatus;
  observacoes?: string;
  createdAt?: string;
  updatedAt?: string;
}