# lib/api-zod/src/generated/types/criarClienteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ContabDOC API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CriarClienteRequestTipo } from "./criarClienteRequestTipo";

export interface CriarClienteRequest {
  escritorioId?: number;
  tipo: CriarClienteRequestTipo;
  cnpj?: string;
  cpf?: string;
  razaoSocial?: string;
  nomeFantasia?: string;
  nomeResponsavel?: string;
  email?: string;
  telefone?: string;
  celular?: string;
  cep?: string;
  logradouro?: string;
  numero?: string;
  complemento?: string;
  bairro?: string;
  municipio?: string;
  uf?: string;
  situacaoReceita?: string;
  socios?: string;
  regimeTributario?: string;
  atividadePrincipal?: string;
}