# lib/api-zod/src/generated/types/criarContratoRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ContabDOC API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CriarContratoRequestStatus } from "./criarContratoRequestStatus";

export interface CriarContratoRequest {
  clienteId: number;
  numeroContrato: string;
  valorContrato?: string;
  dataContrato?: string;
  diaVencimento?: number;
  dataVencimento?: string;
  objeto?: string;
  status?: CriarContratoRequestStatus;
  observacoes?: string;
}