# lib/api-zod/src/generated/types/criarUsuarioRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ContabDOC API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CriarUsuarioRequestPerfil } from "./criarUsuarioRequestPerfil";

export interface CriarUsuarioRequest {
  nome: string;
  email: string;
  senha?: string;
  perfil: CriarUsuarioRequestPerfil;
  ativo?: boolean;
  permissoes?: string;
}