# lib/api-zod/src/generated/types/integracao.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ContabDOC API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Integracao {
  id: number;
  nome: string;
  tipo: string;
  ativo?: boolean;
  config?: string;
  updatedAt?: string;
}