# lib/api-zod/src/generated/types/socio.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ContabDOC API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Socio {
  nome: string;
  qualificacao: string;
  cpfCnpj?: string;
}