# lib/api-zod/src/generated/types/validacaoCpf.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ContabDOC API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ValidacaoCpf {
  cpf: string;
  valido: boolean;
  nome?: string;
  situacao?: string;
}